refactor: move event bus and path helpers to utils
