+++ docs/kernel.md
@@ -7,7 +7,7 @@ The kernel is implemented entirely in TypeScript under `core/`. It maintains the
 - `fs/` – In-memory filesystem. Files are stored as nodes with Unix‑style permissions. Snapshots are persisted via a small SQLite database.
 - `net/` – Simple network stack providing NICs, TCP and UDP sockets. Routers and switches are also written in TypeScript.
 - `services/` – Built‑in daemons such as the HTTP and SSH servers.
-- `eventBus.ts` – Lightweight event emitter used for UI messages and service events.
+- `utils/eventBus.ts` – Lightweight event emitter used for UI messages and service events.
 - `kernel.ts` – The main scheduler, process table and syscall implementations.
 
 Processes are executed inside V8 isolates spawned by the host. Each process has
